requirements.txt
python-binance==1.0.17
requests